Move the optional tests related with storages to the other dir
